# src/docs/apispec.yml
openapi: 3.0.3
info:
  title: Redaktionen API
  description: |
    AI-powered news aggregation and analysis system that performs PESTEL (Political, Economic, Social, Technological, Environmental, Legal) analysis on input signals.

    Developed by Innovationsarenan @ Göteborgsregionen.
  version: 1.0.0
  contact:
    name: Innovationsarenan @ Göteborgsregionen
servers:
  - url: https://redaktionen.innovationsarenan.dev/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication via Authorization header
  schemas:
    Factor:
      type: string
      enum:
        - political
        - economic
        - social
        - technological
        - environmental
        - legal
      description: PESTEL factor category
...
    SourceType:
      type: string
      enum:
        - rss
        - newsletter
      description: Type of news source
    Source:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        agencyId:
          type: string
          description: Associated agency ID
          example: "agency-123"
        source:
          type: string
          description: Name of the source
          example: "BBC News"
        type:
          $ref: "#/components/schemas/SourceType"
        url:
          type: string
          format: uri
          description: URL to the RSS feed or newsletter
          example: "https://feeds.bbci.co.uk/news/rss.xml"
        factor:
          $ref: "#/components/schemas/Factor"
      required:
        - source
        - type
        - url
        - factor
...
    SourcesResponse:
      type: array
      items:
        $ref: "#/components/schemas/Source"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "This is a clear error message"
...
paths:
...
  /sources:
    get:
      tags:
        - Sources
      summary: List sources
      description: Retrieves all news sources (RSS feeds and newsletters) from the database, optionally filtered by PESTEL factor
      operationId: listSources
      security:
        - BearerAuth: []
      parameters:
        - name: factor
          in: query
          description: Filter sources by PESTEL factor
          required: false
          schema:
            $ref: "#/components/schemas/Factor"
          example: environmental
      responses:
        "200":
          description: List of sources retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sources/{sourceId}:
    get:
      tags:
        - Sources
      summary: Get source by ID
      description: Retrieves a single news source by its ID
      operationId: getSource
      security:
        - BearerAuth: []
      parameters:
        - name: sourceId
          in: path
          description: Source ID
          required: true
          schema:
            type: integer
            example: 42
      responses:
        "200":
          description: Source retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Sources
      summary: Delete source
      description: Deletes a news source by its ID
      operationId: deleteSource
      security:
        - BearerAuth: []
      parameters:
        - name: sourceId
          in: path
          description: Source ID
          required: true
          schema:
            type: integer
            example: 42
      responses:
        "200":
          description: Source deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...